Add insane UI/UX Improvements, Fix CSS for SignIns
Added a welcome page that prompts users on website arrival and upon
login. Also fixed up the CSS for the SignIn and SignUp pages to where
they should render properly no matter the window size. Also modified
the backend API to pass back a user's name on successful login, which
also is cached on the users browser as well.